+++ app/Http/Controllers/api/CategoryController.php
@@ -2,7 +2,9 @@
 
 namespace App\Http\Controllers\api;
 
+use Exception;
 use App\Category;
+use App\Interview;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
 use App\Http\Controllers\api\BaseController;
@@ -16,11 +18,11 @@ class CategoryController extends BaseController
      */
     public function index()
     {
-        $category=Category::get();
+        $category=Category::orderBy('id','desc')->paginate(10);
         if (is_null($category)) {
             return $this->sendError('Category not found.');
         }
-        return $this->sendResponse($category->toArray(), 'Category retrieved successfully.');
+        return $this->sendResponse($category, 'Category retrieved successfully.');
     }
 
     /**
@@ -67,23 +69,21 @@ class CategoryController extends BaseController
      * @param  \App\Category  $category
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Category $category)
+    public function update(Request $request, $id)
     {
         $input = $request->all();
+        $category = Category::findOrFail($id);
         $validator = Validator::make($input, [
-            'name' => 'required'
+            'name' => 'required|unique:category,name,'.$id
+        ],
+        [
+            "name.unique"=>"$request->name Email is alerady taken "
         ]);
-        if(category::where('name',$request->name)->count()>0){
-            return $this->sendError('Validation Error.', ['Category Name is already taken']);
-        }
+
         if($validator->fails()){
             return $this->sendError('Validation Error.', $validator->errors());
         }
-        $category->name = $input['name'];
-        $category->hr_id = $input['hr_id'];
-        $category->save();
-
-
+        $category->update($request->all());
         return $this->sendResponse($category->toArray(), 'Category updated successfully.');
     }
 
@@ -95,13 +95,20 @@ class CategoryController extends BaseController
      */
     public function destroy($id)
     {
-        $category=Category::destroy($id);
-        if ($category) {
-            return $this->sendError('Category deleted successfully.');
+        try
+        {
+            if(Interview::where('category_id',$id)->count()>0)
+            {
+                return $this->sendError('Category could not delete candidate is exists.');
+            }
+            $category=Category::destroy($id);
+            return $this->sendResponse($category,'Category deleted successfully.');
         }
-        else
+        catch(Exception $exception)
         {
             return $this->sendResponse($category, 'Category could not delete.');
+
         }
+
     }
 }
